# swagger.yml
swagger: '2.0'
info:
  description: API to manage lessons and their contents.
  version: 1.0.0
  title: Lessons API
  contact: {}
host: localhost:8080
basePath: /
tags:
  - name: Contents Service
    description: Service to manage lesson contents.
  - name: Lessons Service
    description: Service to manage lessons and their contents.
paths:
  /contents:
    get:
      tags:
        - Contents Service
      summary: getContentsBy
      operationId: getContentsByUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: Get all contents
          schema:
            type: array
            items:
              $ref: '#/definitions/ContentDto'
        '204':
          description: No contents found.
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
    post:
      tags:
        - Contents Service
      summary: createContent
      operationId: createContentUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: contentDto
          description: contentDto
          required: true
          schema:
            $ref: '#/definitions/ContentDto'
      responses:
        '201':
          description: Create a new content.
          schema:
            $ref: '#/definitions/ContentDto'
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '404':
          description: Not found route
  /contents/{id}:
    get:
      tags:
        - Contents Service
      summary: getContentById
      operationId: getContentByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Get content by ID
          schema:
            $ref: '#/definitions/ContentDto'
        '204':
          description: Content was not found.
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
    delete:
      tags:
        - Contents Service
      summary: deleteContent
      operationId: deleteContentUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '204':
          description: Delete a content by ID.
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '404':
          description: Not found route
    patch:
      tags:
        - Contents Service
      summary: patchContent
      operationId: patchContentUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: contentDto
          description: contentDto
          required: true
          schema:
            $ref: '#/definitions/ContentDto'
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Patch a content by ID.
          schema:
            $ref: '#/definitions/ContentDto'
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '404':
          description: Not found route
  /lessons:
    get:
      tags:
        - Lessons Service
      summary: getLessonsBy
      operationId: getLessonsByUsingGET
      produces:
        - '*/*'
      parameters:
        - name: classId
          in: query
          description: classId
          required: true
          type: string
      responses:
        '200':
          description: Get all lessons
          schema:
            type: array
            items:
              $ref: '#/definitions/LessonDto'
        '204':
          description: No lessons found.
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
    post:
      tags:
        - Lessons Service
      summary: createLesson
      operationId: createLessonUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: classDto
          description: classDto
          required: true
          schema:
            $ref: '#/definitions/LessonDto'
      responses:
        '201':
          description: Create a new lesson.
          schema:
            $ref: '#/definitions/LessonDto'
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '404':
          description: Not found route
  /lessons/{id}:
    get:
      tags:
        - Lessons Service
      summary: getLessonById
      operationId: getLessonByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Get lesson by ID
          schema:
            $ref: '#/definitions/LessonDto'
        '204':
          description: Lesson was not found.
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
    delete:
      tags:
        - Lessons Service
      summary: deleteLesson
      operationId: deleteLessonUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '204':
          description: Delete a lesson by ID.
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
    patch:
      tags:
        - Lessons Service
      summary: patchLesson
      operationId: patchLessonUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: classDto
          description: classDto
          required: true
          schema:
            $ref: '#/definitions/LessonDto'
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: Patch a lesson by ID.
          schema:
            $ref: '#/definitions/LessonDto'
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
  /lessons/{lessonId}/contents:
    get:
      tags:
        - Lessons Service
      summary: getAllContentsFromLesson
      operationId: getAllContentsFromLessonUsingGET
      produces:
        - '*/*'
      parameters:
        - name: lessonId
          in: path
          description: lessonId
          required: true
          type: string
      responses:
        '200':
          description: Get all contents from lesson.
          schema:
            type: array
            items:
              $ref: '#/definitions/ContentDto'
        '204':
          description: No contents from lesson found.
        '401':
          description: Unauthorized.
        '404':
          description: Not found route.
    post:
      tags:
        - Lessons Service
      summary: createLessonContent
      operationId: createLessonContentUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: contentDto
          description: contentDto
          required: true
          schema:
            $ref: '#/definitions/ContentDto'
        - name: lessonId
          in: path
          description: lessonId
          required: true
          type: string
      responses:
        '201':
          description: Create a new content for a lesson.
          schema:
            $ref: '#/definitions/LessonDto'
        '400':
          description: Business rule error.
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '404':
          description: Not found route
definitions:
  ResponseError:
    type: object
    required:
      - status
      - errors
    properties:
      errors:
        type: object
        example:
          date: "Data obrigatória"
  BuildingDto:
    type: object
    required:
      - id
    properties:
      id:
        type: string
      name:
        type: string
    title: BuildingDto
  ClassDto:
    type: object
    required:
      - id
    properties:
      id:
        type: string
      name:
        type: string
    title: ClassDto
  ContentDto:
    type: object
    required:
      - name
      - type
    properties:
      id:
        type: string
      link:
        type: string
      name:
        type: string
      type:
        type: string
        enum:
          - DOC
          - PDF
          - POWERPOINT
          - SPREADSHEET
    title: ContentDto
  LessonDto:
    type: object
    required:
      - building
      - class
      - date
      - name
    properties:
      building:
        $ref: '#/definitions/BuildingDto'
      class:
        $ref: '#/definitions/ClassDto'
      contents:
        type: array
        items:
          $ref: '#/definitions/ContentDto'
      date:
        type: string
        format: date-time
      id:
        type: string
      name:
        type: string
    title: LessonDto